Fix OOM on write metadata by using one big chunk in template_ds dask arrays (#58)
Don't change the actual encodings of stored data.
Make sure we drive processing based on the encodings, 
not by assuming template dask chunks == encoding.

Dask on the .to_zarr call seems to be trying to cache 
something for every dask chunk which is a massive 
amount (18,726,930 per data var) if dask chunks == chunks 
in our full dataset and causes an OOM.