Don't fail entire shard on assertion errors when downloading and reading specific files
Instead, leave nan in the array. These errors happen with things like malformed files and we don't want to loose out on data from other files in the same shard when an individual file is malformed.

I recently added this in as an extra layer of caution, but in practice would cause unnecessary missing data. We check for missing values both offline in report_nulls and in operational update validators (e.g. check recent nans)